Make schemaMatches more stable
* moved deriveType to util
* schemaMatches doesn't expect a type property but uses deriveType
* added test case for improved schemaMatches behavior
